Major mitoTracing class that store clonal-resolved multi-omics

Major mitoTracing class that store clonal-resolved multi-omics

Slots

GTsummary.filtered

The Mitochondrial genotype data frame

CellMeta

Store meta data for each cell type

V.fitered.list

a list of data frame of variant metrics, VAF, cellN, etc (each for different stringency),

UniqueV

A character showing the number of usable variant

Cts.Mtx

A sparse matrix cell-mitoVariants, store the variant count

Cts.Mtx.bi

A sparse matrix cell-mitoVariants, The variant count has been binarized into 0 and 1

Ctx.Mtx.depth

A sparse matrix cell-mitoVariants(total counts for each position), store the variant count

para

A character showing the parameter of this object

Seurat

Seurat object storing the clonal clustering results

DataToplotList

The customized class of Datatoplots: A list of dataframe for further plotting

DistObjects

The customized class that stores the cell-cell distances

TREE

The customized class that wraps phylogenetic tree