Function to plot variant metrics

This function allows you to plot the mito mutation metrics For each category(stringency), p1: Variant allele frequency(VAF); p2: Heteroplasmy histogram p3: CellN(Number of caells that carry the variants) VS maxcts( The number of variant counts in the highest cell) p4: Histogram to show the distribution of the number of variant per cell

plot_variant(
  GTSummary,
  feature.list,
  depth,
  cat = c("Total", "VerySensitive", "Sensitive", "Specific"),
  p4xlim = 50,
  QualifyCellCut = 10
)

Arguments

GTSummary

The GTSummary file read in by function CW_mgatk.read

feature.list

The variant feature list generated by Vfilter_v3

depth

The .depth file by function DepthSummary

cat

The category(or the striengency to be ploted), default is c("Total","VerySensitive","Sensitive","Specific")

p4xlim

the p4 xlim(number of variant per cell), default is 50

QualifyCellCut

median coverage for qualified cells, default is 10

Value

no returns, directly plot

Examples

plot_variant(DN1CD34_1.VariantsGTSummary,DN1CD34_1.Variants.feature.lst,depth=DN1CD34_1.depth,cat=c("Total","VerySensitive","Sensitive","Specific"),p4xlim = 30)
#> Error in plot_variant(DN1CD34_1.VariantsGTSummary, DN1CD34_1.Variants.feature.lst,     depth = DN1CD34_1.depth, cat = c("Total", "VerySensitive",         "Sensitive", "Specific"), p4xlim = 30): could not find function "plot_variant"